requirements.txt
Django
djangorestframework
gunicorn
whitenoise
psycopg2-binary
django-cors-headers
